requirements.txt
aes-encryption==0.2.0
alembic==1.11.3
annotated-types==0.5.0
blinker==1.6.2
click==8.1.6
exceptiongroup==1.1.2
Flask==2.3.2
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.1
pluggy==1.2.0
protobuf==4.21.12
pycryptodome==3.18.0
pydantic==2.1.1
pydantic_core==2.4.0
pytest==7.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.3.6
